Introducing generalized base query class for app entity queries.
